requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-admin-anchors==2.0.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
drf-flex-fields==0.9.8
drf-spectacular==0.22.1
idna==3.3
inflection==0.5.1
jsonschema==4.6.2
meilisearch==0.18.3
psycopg==3.0.15
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.10
